+++ packages/toolkit/src/index.ts
@@ -1,8 +1,11 @@
 import React from 'react';
 
+import * as gestures from '@use-gesture/react';
+
 export * from './constants';
 export * from './hooks';
 export * from './utils';
+export {gestures};
 import common from './styles/common.module.scss';
 import variables from './styles/variables.css';
 
@@ -32,23 +35,24 @@ export type IColor =
 
 export type ISocials = 'vk' | 'ya' | 'fb' | 'apple';
 
-type ReplaceProps<Component extends React.ElementType, P> = Pick<
+export type ReplaceProps<Component extends React.ElementType, P> = Pick<
   React.ComponentPropsWithRef<Component>,
   Exclude<keyof React.ComponentPropsWithRef<Component>, keyof P>
 > &
   P;
 
 export interface AliasProps<
-  Component extends React.ElementType = React.ElementType
+  Component extends React.ElementType = React.ElementType,
 > {
   as?: Component;
 }
 
-export interface ForwardingComponent<
-  InitialComponent extends React.ElementType,
+export type ForwardingCertainComponent<
+  CertainComponent extends React.ElementType,
+  InitialComponent extends CertainComponent,
   P = unknown
-> {
-  <Component extends React.ElementType = InitialComponent, Context = unknown>(
+> = {
+  <Component extends CertainComponent = InitialComponent, Context = unknown>(
     props: React.PropsWithChildren<
       ReplaceProps<Component, P> & AliasProps<Component>
     >,
@@ -57,3 +61,8 @@ export interface ForwardingComponent<
   defaultProps?: Partial<P>;
   displayName?: string;
 }
+
+export type ForwardingComponent<
+  InitialComponent extends React.ElementType,
+  P = unknown
+> = ForwardingCertainComponent<React.ElementType, InitialComponent, P>;