+++ packages/toolkit/src/utils/index.ts
@@ -6,3 +6,4 @@ export {default as stringifyCssProps} from './stringifyCssProps';
 export {default as smoothScroll} from './smoothScroll';
 export {default as plural} from './plural';
 export {default as prettyNumber} from './prettyNumber';
+export * from './findComponentInChildren';